FizzBuzz
You will write a program in any language of your choosing (please comment the language used at the top of the program) which prints the numbers 1-250, but with the following rules:

If a number is divisible by 3, instead of printing the number, print "Fizz".

If a number is divisible by 5, instead of printing the number, print "Fizzy"

If a number is divisible by 7, instead of printing the number, print "Buzz"

You should print all words in the order "Fizz" then "Fizzy" then "Buzz", if the word is divisible by the appropriate number. For example, 15 should print "FizzFizzy", 21 should print "FizzBuzz", 35 should print "FizzyBuzz", and 105 should print "FizzFizzyBuzz"

Example output sequence for numbers 1-25: 1, 2, Fizz, 4, Fizzy, Fizz, Buzz, 8, Fizz, Fizzy, 11, Fizz, 13, Buzz, FizzFizzy, 16, 17, Fizz, 19, Fizzy, FizzBuzz, 22, 23, Fizz, Fizzy

We are looking for a program which is clean, maintainable and reasonably efficient.

Your GitHub Repository should be public and contain all dependencies.